// FrontEnd/src/app/models/title-number.ts
import { ApplicationForm } from "./application-form";

export class TitleNumber {
    TitleNumberId?: number;
    TitleNumberCode?: string;
    LesseeTitleNumber?: string;
    CreatedDate?: Date;
    UpdatedDate?: Date;

    LocalId?: number;
    IsSelected?: boolean;
    TitleType?: string;

    Status?: boolean;
    DocumentReferenceId?: number;

}